requirements.txt
Naked==0.1.31
PyYAML==3.12
asn1crypto==0.22.0
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
crypto==1.4.1
cryptography==1.9
glop==0.1.1
idna==2.5
pprint==0.1
py-postgresql==1.2.1
pyOpenSSL==17.0.0
pycparser==2.17
pyldap==2.4.35.1
requests==2.17.3
shellescape==3.4.1
six==1.10.0
urllib3==1.21.1
